DRF_NEW/drfecommerce/requirements.txt
asgiref==3.7.2
attrs==23.2.0
colorama==0.4.6
Django==5.0.2
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
